properly encode salt and hash, return new id
This commit properly encodes the generated salt and hashes using the
hex encoder. It also adds the salt and password fields to the user
struct but does not return them or the created/updated at fields in
the JSON response. Lastly the commit includes the real id of the newly
created user in the response.