select user after creation
This commit allows use of structs.Map for the user insert command, which
makes the code more maintainable/readable. Also, we now run a select
command on the newly inserted user in order to get the latest data for
that user. This is necessary because, in the future, we may have hooks
that manipulate the user data during the insertion and we want the
returned user information to be up to date.

We've changed the users schema to use varchars so we don't have to
manually trim whitespace off of the returned column values.