implement user queue update images schema
This commit implements the image queue for a given user. The return
value is a users information along with their first image's info.

This commit also updates the images table schema because the updated_at
and expires_at columns are not necessary. All images are removed after
seven days and they are not updated, just removed by the user.