add access token authentication
This commit changes the API so that an access token is needed for each
endpoint, excluding '/register'. With this new middleware, we can also
parse through a token to receive a users email and, from there, use that
email to find a users id. Since the current users id is no longer
needed, all of the endpoints that once contained the current users' id
have been replaced with 'self' endpoint.

Ex.

/v1/users/{id}/queue has become /v1/users/self/queue

This way, there is no endpoint available to find another users queue
without having their access token, which is nice ;)

Credit: This endpoint design was taken from Instagram's user endpoints.